+++ packages/components/ui/src/gauges/index.js
@@ -1,1 +1,0 @@
-export {default as ScreenGauge} from './screen/ScreenGauge.svelte';