+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - add `NoScript` component
 - add `ResponsiveFlex` component
 - `ScreenGauge`: fixed `sampleText` enlarging the document size
+- rename `ScreenGauge` to `ScreenSensor`
 
 ## `@svizzle/atlas` v0.6.0 (next)